# Add timer to measure performance