--- a/CS3243_P1_XX_Y.py
+++ b/CS3243_P1_XX_Y.py
@@ -1,5 +1,7 @@
 import os
 import sys
+import informed
+import uninformed
 
 
 class Puzzle(object):