- Added eval.py file with helper functions to measure performance.

- Evaluated the performance for 3*3 test inputs

**uninformed search (BFS):**
test_3_1: 0.00001311 s
path length: UNSOLVABLE
test_3_2: 1.90605593 s
path length: 22
test_3_3: 10.53170991 s
path length: 31

**informed search (A star):**
test_3_1: 0.00002098 s
path length: UNSOLVABLE
test_3_2: 0.01765585 s
path length: 22
test_3_3: 0.13642120 s
path length: 31